Bug-fix: Cart page crashing on first load
Bug Cause: On first load of cart page, there was no 'cart' array in the localStorage which was causing the crash.
Solution: Added a method 'createCart', which on load of 'Cart Page' creates and saves an empty 'cart' array to the localStorage if not present already.
This 'createCart' method is executed before loading products in the cart page.