Feature: Added 'Quantity' component in cart product cards
Added showQuantity component method to display quantity info when product is added to cart.
Conditional loading of 'Quantity' and 'Add to Cart' components in cart.
If a product is already available in cart then only it's 'quantity' parameter changes upon pressing 'Add to Cart' button.
Avoids duplication of products in cart.
'+' and '-' buttons on card changes the quantity state of the product and saves it in the localStorage.