Bug-fix: Home Page crashing on first load
Bug Cause: At first load, there is no 'cart' boject present in the localStorge which is required to load the quantity of products in cart and show it in the home page.
Solution: running 'createCart' method at the load of Home page adds an empty cart object in localStorage if not present.