Feature: Added functionality to filter product results.
Settup getProducts method to accept 'queryString' as argument and handle it if argument is present or not.
Created a method 'filter' to make an api call with the queryString to give filter and sort functionality.
Increased the height of the filter card to accomodate the filter submit button.
Added more options in the limit filter for testing.