Added a Dockerfile.dev file to containerize the frontend app.
Used 'npm run start' as a starting/default command to start the app inside the container.
Also linked the volumes inside the container with the our local machine working directory to have realtime file sync.
Using node version 12.18.3 same as my local machine with the alpine version so that the memory footprint is very small.